add apis folders and seprate routes for api endpoints and django functions
