+++ accounts/apis/views.py
@@ -1,11 +1,11 @@
 from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK
 from rest_framework.views import APIView
-from rest_framework.generics import ListAPIView
+from rest_framework.generics import ListAPIView, ListCreateAPIView, RetrieveUpdateDestroyAPIView
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
 from accounts.models import Account
-from accounts.serializers import AccountsListSerializer
+from accounts.serializers import AccountsListSerializer, AccountsCreateSerializer
 
 
 class AccountsListAPIView(APIView):
@@ -28,4 +28,24 @@ class AccountsViewSet(ReadOnlyModelViewSet):
     serializer_class = AccountsListSerializer
 
     def get_queryset(self):
-            return Account.objects.select_related("branch__bank")
+        return Account.objects.select_related("branch__bank")
+
+
+class AccountsCreateListGenericAPIView(ListCreateAPIView):
+    def get_serializer_class(self):
+        if self.request.method == "POST":
+            return AccountsCreateSerializer
+        return AccountsListSerializer
+
+    def get_queryset(self):
+        return Account.objects.select_related("branch__bank")
+
+
+class AccountsRetrieveUpdateDestroyAPIView(RetrieveUpdateDestroyAPIView):
+    def get_serializer_class(self):
+        if self.request.method == "GET":
+            return AccountsListSerializer
+        return AccountsCreateSerializer
+
+    def get_queryset(self):
+        return Account.objects.select_related("branch__bank")
