+++ accounts/views.py
@@ -2,15 +2,8 @@ from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
 from django.views import View
 from django.utils.decorators import method_decorator
-from rest_framework.response import Response
-from rest_framework.status import HTTP_200_OK
-from rest_framework.views import APIView
-from rest_framework.generics import ListAPIView, ListCreateAPIView, RetrieveUpdateDestroyAPIView
-from rest_framework.viewsets import ReadOnlyModelViewSet
 
 from accounts.models import Account
-from accounts.serializers import AccountsListSerializer, AccountsCreateSerializer
-
 
 
 @method_decorator(login_required(login_url="/users/login/"), name="dispatch")
@@ -24,52 +17,3 @@ class AccountsListView(View):
             } for acc in accounts]
             
             return JsonResponse({"data": data})
-
-
-class AccountsListAPIView(APIView):
-    def get(self, request, *args, **kwargs):
-        accounts = Account.objects.select_related("branch__bank").filter(user=self.request.user)
-        serializer = AccountsListSerializer(accounts, many=True)
-        context = {"data": serializer.data, "count": accounts.count()}
-
-        return Response(context, status=HTTP_200_OK)
-
-
-class AccountsListGenericAPIView(ListAPIView):
-    serializer_class = AccountsListSerializer
-
-    def get_queryset(self):
-        return Account.objects.select_related("branch__bank").filter(user=self.request.user)
-
-
-class AccountsViewSet(ReadOnlyModelViewSet):
-    serializer_class = AccountsListSerializer
-
-    def get_queryset(self):
-        if self.request.user.is_authenticated:
-            return Account.objects.select_related("branch__bank").filter(user=self.request.user)
-        return Account.objects.none()
-
-
-class AccountsCreateListGenericAPIView(ListCreateAPIView):
-    def get_serializer_class(self):
-        if self.request.method == "POST":
-            return AccountsCreateSerializer
-        return AccountsListSerializer
-
-    def get_queryset(self):
-        if self.request.user.is_authenticated:
-            return Account.objects.select_related("branch__bank").filter(user=self.request.user)
-        return Account.objects.none()
-
-
-class AccountsRetrieveUpdateDestroyAPIView(RetrieveUpdateDestroyAPIView):
-    def get_serializer_class(self):
-        if self.request.method == "GET":
-            return AccountsListSerializer
-        return AccountsCreateSerializer
-
-    def get_queryset(self):
-        if self.request.user.is_authenticated:
-            return Account.objects.select_related("branch__bank").filter(user=self.request.user)
-        return Account.objects.none()
