+++ banks/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from banks.views import BankListCreateGenericView, BankRetrieveUpdateDestroyGenericView
+from banks.views import BankListView
 
 urlpatterns = [
-    path("", BankListCreateGenericView.as_view(), name="bank-list-create"),
-    path("<int:pk>/", BankRetrieveUpdateDestroyGenericView.as_view(), name="bank-detail"),
+    path("", BankListView.as_view(), name="bank-list"),
 ]
