+++ banks/views.py
@@ -3,14 +3,8 @@ from django.db.models import Count
 from django.http import JsonResponse
 from django.utils.decorators import method_decorator
 from django.views import View
-from rest_framework.status import HTTP_200_OK
-from rest_framework.views import APIView
-from rest_framework.generics import ListAPIView, ListCreateAPIView, RetrieveUpdateDestroyAPIView
-from rest_framework.viewsets import ReadOnlyModelViewSet
-from rest_framework.response import Response
 
 from banks.models import Bank
-from banks.serializers import BankListSerializer, BankCreateSerializer
 
 
 @method_decorator(login_required(login_url="/users/login/"), name="dispatch")
@@ -29,53 +23,3 @@ class BankListView(View):
         ]
 
         return JsonResponse({"data": data})
-
-
-class BankListApiView(APIView):
-    def get(self, request, *args, **kwargs):
-        banks = Bank.objects.annotate(
-            branch_count=Count("branches")
-        )
-        serializer = BankListSerializer(banks, many=True)
-        context = {"data": serializer.data, "count": banks.count()}
-
-        return Response(context, status=HTTP_200_OK)
-
-
-class BankListGenericView(ListAPIView):
-    serializer_class = BankListSerializer
-    
-    def get_queryset(self):
-        return Bank.objects.annotate(
-            branch_count=Count("branches")
-        )
-
-
-class BankViewSet(ReadOnlyModelViewSet):
-    serializer_class = BankListSerializer
-    
-    def get_queryset(self):
-        return Bank.objects.annotate(
-            branch_count=Count("branches")
-        )
-
-
-class BankListCreateGenericView(ListCreateAPIView):
-    def get_serializer_class(self):
-        if self.request.method == 'POST':
-            return BankCreateSerializer
-        return BankListSerializer
-
-    def get_queryset(self):
-        return Bank.objects.annotate(
-            branch_count=Count("branches")
-        )
-
-
-class BankRetrieveUpdateDestroyGenericView(RetrieveUpdateDestroyAPIView):
-    queryset = Bank.objects.all()
-    
-    def get_serializer_class(self):
-        if self.request.method in ['PUT', 'PATCH']:
-            return BankCreateSerializer
-        return BankListSerializer
